Add files via upload
This exercise deals with the story of the emperor who thanked the inventor of chess and asked him to name a reward. The inventor asked for one grain on the first square, 2 for the second, 4 for the third and so forth. The program will determine and display when the grains reach 1,000, 1,000,000 and 1,000,000,000 grains.